requirements.txt
Babel==0.9.6
Beaker==1.5.4
BeautifulSoup==3.2.1
Django==1.4
FormEncode==1.2.3dev
Genshi==0.6
Mako==0.3.6
Markdown==2.0.3
MarkupSafe==0.11
PIL==1.1.7
Paste==1.7.4
PasteDeploy==1.3.3
PasteScript==1.7.3
Pygments==1.4
Pylons==0.10
Routes==1.12.3
SQLAlchemy==0.6.5
South==0.7.2
Tempita==0.5dev
ToscaWidgets==0.9.9
Unidecode==0.04.5
-e git://github.com/chrisdickinson/waiter.git@7a23eea089ed5f63fcb404514b780217e854d749#egg=Waiter-dev
WebError==0.10.3
WebHelpers==1.0
WebOb==1.0.7
WebTest==1.2
akismet==0.2.0
amqplib==0.6.1
anyjson==0.2.4
carrot==0.10.5
celery==2.0.3
-e git://github.com/chrisdickinson/d51_admin_autofk.git@5077973cf829bc3851daa063b337bfbc69b35389#egg=d51_admin_autofk-dev
-e git://github.com/chrisdickinson/d51_django_admin_piston@fff19481c287b9db51db9166882e0b09e1622764#egg=d51_django_admin_piston-dev
decorator==3.3.3
django-celery==2.0.2
django-grappelli==2.3.9
django-picklefield==0.1.6
-e hg+http://bitbucket.org/jespern/django-piston/@e539a104d5165e5a1d5ac72c0c06a4d7cf235c32#egg=django_piston-dev
django-tinymce==1.5.1
feedparser==4.1
gdata==2.0.17
gravatar==0.1
gunicorn==0.14.6
httplib2==0.7.6
nose==1.0.0
psycopg2==2.0.8
pyparsing==1.5.5
python-dateutil==1.5
simplejson==2.1.3
sorl-thumbnail==11.12
sqlalchemy-migrate==0.6
-e git://github.com/chrisdickinson/surlex.git@2b40e4f9a0e7455b911b767a575bf21330d82a94#egg=surlex-dev
-e git://github.com/chrisdickinson/tag_utils.git@0050cdf02313fd964a8702a9224bdf6a25942381#egg=tag_utils-dev
tw.forms==0.9.9
virtualenv==1.3.3
wsgiref==0.1.2
zope.interface==3.6.1
